claim changes

+++ src/main/java/com/example/claim/Claim.java
@@ -23,8 +23,8 @@ public class Claim {
 
     private String claimNumber;
 
-    private Integer requestedAmount;
-    private Integer approvedAmount;
+    private BigDecimal requestedAmount;
+    private BigDecimal approvedAmount;
 
     private String description;
