+++ src/main/java/com/example/claim/dto/UpdateClaimRequest.java
@@ -1,5 +1,7 @@
 package com.example.claim.dto;
 
+import java.math.BigDecimal;
+
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.PositiveOrZero;
 import lombok.AllArgsConstructor;
@@ -18,5 +20,5 @@ public class UpdateClaimRequest {
     private String remarks; // Optional by default
 
     @PositiveOrZero(message = "Approved amount cannot be negative")
-    private Integer approvedAmount;
+    private BigDecimal approvedAmount;
 }
